Remove build target for rustc 1.0.0, add for 1.2.0.
It was a hope, but we shouldn't try to build for rustc versions earlier than what's in use now.
Instead, we pin 1.2.0 to ensure that even once stable moves forward, future commits
don't break changes to the oldest supported version.

So why 1.2.0? We use std::iter::once, which was introduced in that version.
If this commit builds, that becomes our Official Oldest Supported Version--right up until
we need something from a newer release. At least we'll definitely know that the build broke
something, at least.